task_manager/requirements.txt
Django==5.0
mysqlclient==2.2.1
python-dotenv==1.0.0
django-environ==0.11.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0